+++ R/gg-constructors.R
@@ -186,9 +186,10 @@ agg_label <- function(text, x, y, panel, colour = "black", size = 20, color) {
 #' Add automatically placed label
 #'
 #' @param quiet (logical, default FALSE) Do not output progress of autolabeller
+#' @param arrow_lines (logical, default TRUE) Add arrows between labels and line series
+#' if there isn't line of sight between the label and the series.
 #' @param arrow_bars (logical, default FALSE) Add arrows between labels and bar series
-#' if there isn't line of sight between the label and the series. Line series always
-#' have arrows (if there isn't line of sight).
+#' if there isn't line of sight between the label and the series.
 #'
 #' @seealso \code{vignette("plotting-options", package = "arphit")} for a detailed description of
 #' all the plotting options
@@ -201,8 +202,8 @@ agg_label <- function(text, x, y, panel, colour = "black", size = 20, color) {
 #'   agg_autolabel()
 #'
 #' @export
-agg_autolabel <- function(quiet = FALSE, arrow_bars = FALSE) {
-  return(list(type = "autolabel", quiet = quiet, arrow_bars = arrow_bars))
+agg_autolabel <- function(quiet = FALSE, arrow_lines = TRUE, arrow_bars = FALSE) {
+  return(list(type = "autolabel", quiet = quiet, arrow_lines = arrow_lines, arrow_bars = arrow_bars))
 }