+++ NEWS.md
@@ -7,9 +7,10 @@
  for the right hand side. Vertically divided time series graphs always add a margin. 
  ([#292](https://github.com/angusmoore/arphit/pull/292))
  * Export a list of graphs as a GIF animation ([#293](https://github.com/angusmoore/arphit/pull/293))
+ * Option to stop autolabeller adding arrows for line series ([#302](https://github.com/angusmoore/arphit/pull/302))
  * On-panel legends. Optional placement arguments in `agg_legend` to specify where
  to place on-panel legends. By default, legends are beneath the graph. ([#304](https://github.com/angusmoore/pull/304))
-
+ 
 ## Breaking changes
 
 ## Bugfixes